Refactor and Decoupling PDFProcessor
