+++ app/src/main/res/values/banking_group_array.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="banking_group_list">
+        <item>HSBC</item>
+
+    </string-array>
+</resources>
